requirements.txt
# requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
python-dotenv>=1.0.0
httpx>=0.28.0

# AI/ML dependencies
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-experimental>=0.0.50
langgraph>=0.0.60
pydantic>=2.0.0

# Data processing
openpyxl>=3.1.0
xlrd>=2.0.0
chardet>=5.0.0

# Optional utilities
numpy>=1.24.0
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-docx>=0.8.11

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0